OpenGL Context

+++ Engine/src/Layers/ImGuiLayer.cpp
@@ -61,7 +61,7 @@ namespace ENGINE
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
@@ -81,7 +81,6 @@ namespace ENGINE
 		ImGui_ImplOpenGL3_Shutdown();
 		ImGui_ImplGlfw_Shutdown();
 		ImGui::DestroyContext();
-
 	}
 
 	void ImGuiLayer::OnEvent(Event& event)
